backend/requirements.txt
fastapi
uvicorn
pydantic
playwright
google-genai
openai # 新增
python-dotenv
httpx
keyring
